[sre] Separate SRE MonoDynamicImage encoding and SRE Save functions to separate files. NFC
- (sre-encode.c) Encoding SRE builders into a MonoDynamicImage and token creation.

- (sre-save.c) Everything for saving SRE images to disk.
